Removed prototyes for static inline routines and moved the comments into the inline implementation.   The impetus for this was twofold.  First, it is incorrect to have static inline prototypes when using the macro implementation.  Second, this reduced the number of lines in the include files seen by rtems.h by about 2000 lines.
Next we restricted visibility for the inline routines to inside the
executive itself EXCEPT for a handful of objects.  This reduced the
number of include files included by rtems.h by 40 files and reduced
the lines in the include files seen by rtems.h by about 6000 lines.

In total, these reduced the compile time of the entire RTEMS tree by 20%.
This results in about 8 minutes savings on the SparcStation 10 morgana.