Switched from driver's own buffering to using termios.  This eliminated the need for the ringbuffer manipulation in this driver.  It should also allow the removal of mc68681_flush.  This greatly simplified the interrupt processing.
Reworked Interrupt Mask Register handling.  Added imr field to the device
context.  This handling makes sure that the IMR accurately reflects the
state of the port in use and its mate.  This required the reworking of
the mc68681_enable_interrupts() routine to allow for the specification
of which interrupt sources were being enabled/disabled.

Reworked initialization to reset using port address rather than chip
base address.  Basically we were always resetting port 0.

Added deviceType field which allows for multiple chips of the same
class to have the same vector number.