DCR setting changed to match what the DY-4 Firmware initialized it to. This primarily included setting the state of the programmable outputs and the RX and TX FIFO depths.
Moved all of the TX, RX, and RRA data structure initialization to before
the hardware initialization.  As part of this, the hardware initialization
was consolidated.  More than likely, some of this movement broke stuff.

Used constants added to sonic.h which gave more logical names to some
of the register bit settings.

Switched to calloc to insure the data areas where initialized to 0.

Commented out a panic check in the RX server which may or may not have
been right.

Increased the size of the CAM initialization area.  It is possible
that this could be decreased or code added to handle the management
of multiple hardware addresses.

Added sonic read and write register routines which aid greatly in
debugging and provide the core for the eventual movement of this
driver to libchip.

Added debug code to the read and write register routines which can
print the value read from or written to a register.  This code also
prints the register name which significantly eases reading the log.