Working on being able to cut RPMs.  It now appears that the process is working mostly for sparc at least.
There is one major problem -- the current process generates a unique
source file per RPM when in fact all of the source files are nearly
the same -- it is only the spec part of the rpm which differs.
The new file mkbinutils_subpackage_version is an attempt to
address this.  It does part of the job right -- one source file
produces multiple binary RPMs.  BUT the end user can not produce
the resulting RPMS themselves from SRPMS unless they also
build all targets.