Significantly lowered the default memory requirements:     - CONFIGURE_RTEMS_INIT_TASKS_TABLE was 10 now 0     - CONFIGURE_POSIX_INIT_THREAD_TABLE was 10 now 0     - CONFIGURE_ITRON_INIT_TASK_TABLE was 10 now 0     - CONFIGURE_LIBIO_MAXIMUM_FILE_DESCRIPTORS was 20 now 3     - added CONFIGURE_NUMBER_OF_TERMIOS_PORTS and defaulted to 1     - added CONFIGURE_TERMIOS_DISABLED defaulted to "enabled"     - miniIMFS is now the default
Added configuration error checks that:
  + Ensure > 0 tasks/threads are configured
  + Ensure at least one inititalization task/thread is defined

bsp.h now defines these so BSP specific requirements
are accounted for.
  + CONFIGURE_NUMBER_OF_TERMIOS_PORTS
  + CONFIGURE_INTERRUPT_STACK_MEMORY

console_reserve_resources and rtems_termios_reserve_resources
are no longer required and considered obsolete.  Calls to
rtems_termios_reserve_resources have been eliminated although
the routine is still there and the body "if 0'ed".

We are very close to having NO reason to modify the
configuration tables in the BSP.  Be warned that eventually
we would like to see the need for BSP_Configuration
eliminated!