Update from Stephan Wilms <Stephan.Wilms@CWA.de> that adds autodetection.
What I implemented in this new driver version is, that the driver will
first probe for a DEC21140 card and use it if found. If not found it
will probe for a DEC21143 card and use that if found. This removes the
need for defining/undefining a macro (as was required with my previous
version). I tested the driver with my 21143 card using netdemo and it
apears to work just fine.

Here are some cases that I did not test, mainly because I do not have
the required testing hardware:
   - I did not enhance or test PPC support
   - I did not test DEC21140 support
   - I did not use other test software than netdemo