Milkymist - Increase SOFTUSB_PMEM_SIZE to 8 KB
We increased PMEM to 8 kB but never updated SOFTUSB_PMEM_SIZE.
This didn't have much of an effect because all we used it for was
to zero unused program memory - which we should never touch
anyway.

But now that "usb load" in FN uses SOFTUSB_PMEM_SIZE, this caused
it to silently truncate the firmware, producing rather puzzling
effects.

Signed-off-by: Joel Sherrill <joel.sherrill@oarcorp.com>