score: Fix performance issue for 64-bit timestamps
The 64-bit timestamps were introduced to simplify the timestamp
calculations.  This works well since nearly all operations are
additions.  The previous _TOD_Tickle_ticks() implementation had a
serious performance regression in case of 64-bit timestamps due to the
usage of two 64-bit divisions which are quite expensive on some
architectures.

A new field seconds_trigger in TOD_Control is introduced to trigger the
_Watchdog_Tickle_seconds() in _TOD_Tickle_ticks().  This avoids the
64-bit divisions completely and only 32-bit additions are used.