libblock: Fix read-ahead task wake-up
In case the read-ahead task is not configured we must not send a wake-up
event.  This would send the wake-up event to the executing task.

We must send the wake-up event only in case the request list changes
from empty to non-empty.  Since otherwise we may send a false transfer
event.