score: PR1607: Add and use CPU_SIZEOF_POINTER
Add and use new CPU port define CPU_SIZEOF_POINTER.  It must be an
integer literal that can be used by the assembler.  This value will be
used to calculate offsets of structure members.  These offsets will be
used in assembler code.

The size of a pointer is part of the application binary interface (ABI)
and thus independent of the actual programming language.  The compiler
will provide defines to determine the current ABI.  We use these defines
to select the appropriate CPU_SIZEOF_POINTER value.

Static assertions in the new file "cpukit/score/src/percpuasm.c" will
ensure that the value of CPU_SIZEOF_POINTER is consistent with the
current compiler settings.  Also the offset values used by assembler
code are verfied.