Restored to pre PPC440 changes.
This file was originally modified by me to support the PPC440 of the virtex5
BSP.  I have since discovered that the clock driver built with the MVME3100
BSP works out of the box for the virtex5 and so have switched to using that.
This file is reverted to the way it was before I touched it, modulo whitespace,
making some functions static (to avoid compiler warnings) and changes by others.