bsps/arm: Add PHY up/down
All PHYs will be set to power up/down state with the BMCR[BMCR_PDOWN]
setting.

The KSZ80X1RNL will use the energy detect power down (EDPD) mode with
PLL automatically turned in the up state.  It will use the slow
oscillator mode in the down state.

To prevent system bus lock-ups the PHY is set to the up state before
Ethernet module initialization.  In case of a communication failure with
the PHY or weird PHY identifiers the initialization will be aborted.