doc: Support texi2any and texi2html
This was tested with texi2html-1.82-5.1.el6.noarch and
a locally built texinfo 5.0. These are completely different
implementations and require different invocations.

The Makefile dependencies appear to work but are likely not
perfect at this point.  The key point is that the autoconf
probe detects which to use and responds accordingly with
preference given to texi2any.