doc/user Object Id Figures: Correct rotation and border
The 16 bit Object Id image was incorrectly rotated in the manual.
Both the 16 and 32 Object Id bit images had thin borders on the table
and the bottom line was not visible in the PDF. I tinkered with the
original drawings to add a bit more space around the tables and
manually converted the PNG to EPS using Gimp to ensure a controlled
and high quality conversion.