Add a new necessary definition needed for raspberrypi MMU support
The new ARM_CP15_CTRL_XP is necessary to share ARMv6 and ARMv7
page-table formats and definitions.
It enables the extended page tables (introduced in ARMv6)
to be configured for the hardware page translation mechanism. This way
we can share ARMv6 and ARMv7 page tables entry formats.

Other Fault Status Register Definitions can be useful for debugging or
excpetion handlers.