PR2157: fdisk partition table dump improvement
Show the correct index of partition's last block (partition end).

The documentation of struct rtems_bdpart_partition (P) says that the member
'end' is the "Block index for partition end (this block is not a part of the
partition)". Then, the fdisk's partition table dump should print ((P)->end -
1).

Currently, one can think that the last block of a partition P is superposing
the beginning of the partition (P + 1). Example:

----------------------------------------
       PARTITION TABLE
------------+------------+--------------
 BEGIN      | END        | TYPE
------------+------------+--------------
       2048 |     133120 |       FAT 32
     133120 |   15628032 |       FAT 32
------------+------------+--------------

With be proposed patch, it would be:
----------------------------------------
       PARTITION TABLE
------------+------------+--------------
 BEGIN      | END        | TYPE
------------+------------+--------------
       2048 |     133119 |       FAT 32
     133120 |   15628031 |       FAT 32
------------+------------+--------------