tools: Add rtems-bsp to list BSPs in the source tree.
This is a shell script that lists the BSPs in the source tree by
architecture. The BSP name is given plus the relative path to
BSP source files.

The script uses the same hack, ie looking for *.cfg files, as the
build system so it sees what the build system sees.