bsp/ngmp: Use -mcpu=leon3 GCC option
There is support for the LEON3 processor available in Binutils 2.24 and
the GCC 4.8 branch and GCC mainline.

GCC 4.8 branch:

http://gcc.gnu.org/viewcvs/gcc/branches/?view=log&pathrev=205331

GCC mainline:

http://gcc.gnu.org/viewcvs/gcc/trunk/?view=log&pathrev=202664

It is mandatory to use this option for SMP on LEON3 since it enables
usage of C11 atomic operations.  It makes it also possible to use an
inline function for _CPU_SMP_Get_current_processor() which avoids the
function call overhead in critical sections.