preliminary Raspberry Pi Model 2 support
This patch adds a BSP variant for the Raspberry Pi 2. You can
build both variants by configuring with the option
 --enable-rtemsbsp="raspberrypi2 raspberrypi"

For the current BSP, the only change was the peripheral register base
address and the compiler options.

The raspberrypi/make/custom rules were re-factored:
  raspberrypi.inc  -- Common rules
  raspberrypi.cfg  -- Raspberry Pi 1 specific rule/optons
  raspberrypi2.cfg -- Raspberry Pi 2 specific rule/options

I tested hello, ticker, unlimited, and paranoia on both the Pi (Model A+)
and Pi 2.